split memory storage to message and memory
